requirements.txt
numpy
opencv-python
pyrender
trimesh
pillow
ultralytics
